tests/requirements.txt
-r ../requirements.txt

boto
boto3
botocore
flake8
mock
moto
pytest
pytest-helpers-namespace
pytest-salt
